services/analytics-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
apscheduler==3.10.4
pydantic==2.7.1
prometheus-client==0.20.0
python-dotenv==1.0.1

